requirements.txt
pyinstaller
paramiko
PyYAML
wget
timeout_decorator